add new sensor requests, switch to mock client with test data

+++ src/FronApiUs.Core/Models/CommonInverterData.cs
@@ -27,7 +27,7 @@ public class CommonInverterDataContent
     public Quantity<decimal> VoltageDC { get; set; } = new();
 
     [JsonPropertyName("DAY_ENERGY")]
-    public Quantity<int> EnergyToday { get; set; } = new();
+    public Quantity<decimal> EnergyToday { get; set; } = new();
 
     [JsonPropertyName("YEAR_ENERGY")]
     public Quantity<decimal> EnergyThisYear { get; set; } = new();
